# services/payments-svc/internal/server/openapi.yaml
...
openapi: 3.0.3
info:
    title: Omnibus network payment hub API
    version: 1.0.0
paths:
...
    /v1/operator/clock:
        post:
            tags:
                - NetworkOperationsService
            description: |-
                SetClock moves the scenario clock (and the chain's) forward. Demo builds
                 only; a production build answers PERMISSION_DENIED.
            operationId: NetworkOperationsService_SetClock
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.SetClockRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.SetClockResponse'
...
    /v1/ops/defunds:
        post:
            tags:
                - BankOperationsService
            description: RequestDefund asks for free position back to the master account (maker).
            operationId: BankOperationsService_RequestDefund
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.RequestDefundRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Defund'
    /v1/ops/defunds/{id}/approve:
        post:
            tags:
                - BankOperationsService
            description: ApproveDefund is the checker; never the requester, always the same bank.
            operationId: BankOperationsService_ApproveDefund
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.ApproveDefundRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Defund'
    /v1/ops/funding:
        post:
            tags:
                - BankOperationsService
            description: |-
                FundOmnibus sends reserves into the joint account: Fedwire BTRC while
                 Fedwire is open, FedNow LMT1 while it is closed.
            operationId: BankOperationsService_FundOmnibus
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.FundOmnibusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.FundOmnibusResponse'
...
    /v1/ops/holds/{id}/block:
        post:
            tags:
                - BankOperationsService
            description: |-
                BlockHold: a sanctioned party has an interest in the funds. They move to
                 the bank's blocked account (BLCK) and a report is due to OFAC within 10
                 business days (31 CFR 501.603).
            operationId: BankOperationsService_BlockHold
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.HoldDecisionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/ops/holds/{id}/reject:
        post:
            tags:
                - BankOperationsService
            description: |-
                RejectHold: prohibited, but no sanctioned party has an interest. Nothing
                 moves (RR04); reported within 10 business days (31 CFR 501.604).
            operationId: BankOperationsService_RejectHold
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.HoldDecisionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/ops/holds/{id}/release:
        post:
            tags:
                - BankOperationsService
            description: ReleaseHold clears a false positive; the payment proceeds.
            operationId: BankOperationsService_ReleaseHold
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.HoldDecisionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
...
    /v1/ops/payments:
        get:
            tags:
                - BankOperationsService
            description: ListBankPayments lists the bank's clients' payments, with staff detail.
            operationId: BankOperationsService_ListBankPayments
            parameters:
                - name: status
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.ListBankPaymentsResponse'
...
    /v1/payees/verify:
        post:
            tags:
                - PaymentService
            description: |-
                VerifyPayee asks the receiving bank whether an account is open and the
                 name matches its records.
            operationId: PaymentService_VerifyPayee
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.Party'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.PayeeCheck'
...
    /v1/payments:
        get:
            tags:
                - PaymentService
            description: ListPayments lists the caller's organization's payments, newest first.
            operationId: PaymentService_ListPayments
            parameters:
                - name: status
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.ListPaymentsResponse'
        post:
            tags:
                - PaymentService
            description: CreatePayment records a payment order from a maker.
            operationId: PaymentService_CreatePayment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.CreatePaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.CreatePaymentResponse'
    /v1/payments/{id}:
        get:
            tags:
                - PaymentService
            description: GetPayment returns one payment with its audit trail.
            operationId: PaymentService_GetPayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payments/{id}/approve:
        post:
            tags:
                - PaymentService
            description: |-
                ApprovePayment records an approval. The maker can never approve its own
                 payment; each approver counts once.
            operationId: PaymentService_ApprovePayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.ApprovePaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payments/{id}/cancel:
        post:
            tags:
                - PaymentService
            description: |-
                CancelPayment withdraws a payment not yet approved (its maker), or pulls
                 a NORMAL payment from the netting queue before its cycle runs.
            operationId: PaymentService_CancelPayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.CancelPaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
    /v1/payments/{id}/decline:
        post:
            tags:
                - PaymentService
            description: DeclinePayment is an approver refusing a payment.
            operationId: PaymentService_DeclinePayment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.DeclinePaymentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Payment'
...
    /v1/webhooks:
        get:
            tags:
                - WebhookService
            description: ListWebhooks lists the organization's endpoints, without secrets.
            operationId: WebhookService_ListWebhooks
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.ListWebhooksResponse'
        post:
            tags:
                - WebhookService
            description: CreateWebhook registers an https endpoint; the secret is returned once.
            operationId: WebhookService_CreateWebhook
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.CreateWebhookRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Webhook'
    /v1/webhooks/{id}:
        delete:
            tags:
                - WebhookService
            description: DeleteWebhook removes an endpoint; its pending deliveries are dropped.
            operationId: WebhookService_DeleteWebhook
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.DeleteWebhookResponse'
...
components:
    schemas:
...
        api.payments.v1.ApproveDefundRequest:
            type: object
            properties:
                id:
                    type: string
        api.payments.v1.ApprovePaymentRequest:
            type: object
            properties:
                id:
                    type: string
        api.payments.v1.CancelPaymentRequest:
            type: object
            properties:
                id:
                    type: string
        api.payments.v1.CreatePaymentRequest:
            type: object
            properties:
                debtorAccount:
                    type: string
                creditor:
                    $ref: '#/components/schemas/api.payments.v1.Party'
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                priority:
                    type: string
                remittance:
                    type: string
                endToEndId:
                    type: string
        api.payments.v1.CreatePaymentResponse:
            type: object
            properties:
                payment:
                    $ref: '#/components/schemas/api.payments.v1.Payment'
                duplicateRequest:
                    type: boolean
                    description: |-
                        True when the Idempotency-Key matched an earlier request with the same
                         body: nothing new was created.
        api.payments.v1.CreateWebhookRequest:
            type: object
            properties:
                url:
                    type: string
        api.payments.v1.DeclinePaymentRequest:
            type: object
            properties:
                id:
                    type: string
                reason:
                    type: string
        api.payments.v1.Defund:
            type: object
            properties:
                id:
                    type: string
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                requestedBy:
                    type: string
                approvedBy:
                    type: string
                status:
                    type: string
                createdAt:
                    type: string
                    format: date-time
...
        api.payments.v1.FundOmnibusRequest:
            type: object
            properties:
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
        api.payments.v1.FundOmnibusResponse:
            type: object
            properties:
                instrument:
                    type: string
                status:
                    type: string
                reason:
                    type: string
                paymentsReleased:
                    type: integer
                    format: int32
        api.payments.v1.HoldDecisionRequest:
            type: object
            properties:
                id:
                    type: string
                note:
                    type: string
...
        api.payments.v1.ListBankPaymentsResponse:
            type: object
            properties:
                payments:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.Payment'
                nextPageToken:
                    type: string
...
        api.payments.v1.ListPaymentsResponse:
            type: object
            properties:
                payments:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.Payment'
                nextPageToken:
                    type: string
...
        api.payments.v1.Party:
            type: object
            properties:
                name:
                    type: string
                routingNumber:
                    type: string
                account:
                    type: string
            description: Party identifies a payee by name, ABA routing number and account.
        api.payments.v1.PayeeCheck:
            type: object
            properties:
                accountStatus:
                    type: string
                nameMatch:
                    type: string
                registeredName:
                    type: string
            description: |-
                PayeeCheck is the receiving bank's answer before any money moves, as the
                 Fed's Payee Name Verification and banks' account-validation services give
                 it.
        api.payments.v1.Payment:
            type: object
            properties:
                id:
                    type: string
                endToEndId:
                    type: string
                uetr:
                    type: string
                idempotencyKey:
                    type: string
                organizationId:
                    type: string
                debtorAccount:
                    type: string
                creditor:
                    $ref: '#/components/schemas/api.payments.v1.Party'
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                priority:
                    type: string
                remittance:
                    type: string
                status:
                    type: string
                    description: |-
                        AWAITING_APPROVAL | IN_PROCESS | ON_HOLD | AWAITING_LIQUIDITY |
                         PENDING_RECEIVER | QUEUED_FOR_NETTING | SETTLED | REJECTED | CANCELLED |
                         EXPIRED | BLOCKED
                isoStatus:
                    type: string
                    description: RCVD | ACTC | PATC | ACCP | PDNG | ACSP | ACSC | ACCC | RJCT | CANC | BLCK
                reasonCode:
                    type: string
                route:
                    type: string
                    description: on-network instant, 24x7 | network netting cycle | book transfer, same bank
                routeNote:
                    type: string
                payeeCheck:
                    $ref: '#/components/schemas/api.payments.v1.PayeeCheck'
                approvalsRequired:
                    type: integer
                    format: int32
                approvals:
                    type: array
                    items:
                        type: string
                createdBy:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                settledAt:
                    type: string
                    format: date-time
                ofacReportDue:
                    type: string
                    format: date-time
                history:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.payments.v1.PaymentEvent'
                fileId:
                    type: string
                sanctionsMatch:
                    type: string
            description: Payment is a payment order and its lifecycle.
...
        api.payments.v1.RequestDefundRequest:
            type: object
            properties:
                amount:
                    $ref: '#/components/schemas/api.payments.v1.Money'
...
        api.payments.v1.SetClockRequest:
            type: object
            properties:
                time:
                    type: string
                    format: date-time
        api.payments.v1.SetClockResponse:
            type: object
            properties:
                now:
                    type: string
                    format: date-time
...